Reorganizing and a bit of New Lore
Remembered to update the revision number this time! Also took 1336 off
of migrant in charge of formatting since he's kind of distanced himself
from the Project :(

I might delete the LFE News section if nobody's gonna maintain it, or at
the least add a disclaimer that it's not gonna be updated anymore.

I still need to go through and start grouping everything logically
again.